+++ test/fixtures/dts-before/vue-component.d.ts
@@ -1,0 +1,26 @@
+import type { DefineComponent, Ref } from "vue";
+
+export declare const TExample: DefineComponent<
+  { value: string },
+  { focus: () => void },
+  {},
+  {},
+  {},
+  {},
+  {},
+  {},
+  string,
+  import("vue").PublicProps,
+  Readonly<{ value: string }>,
+  {},
+  {},
+  {},
+  string,
+  {},
+  {},
+  {},
+  {},
+  {}
+>;
+
+export declare const count: Ref<number, number>;